# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A simple Order tracking API for an ecommerce site. View source code here: https://github.com/grqphical07/order-api",
        "title": "Order API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/grqphical07/order-api/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "summary": "Base Route",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderapi.IndexResponse"
                        }
                    }
                }
            }
        },
        "/add-order": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Adds an order to the system",
                "parameters": [
                    {
                        "description": "Order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/orderapi.Order"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/orderapi.Order"
                        }
                    },
                    "500": {
                        "description": "Failed to parse JSON",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/complete-order": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Deactivates an order and archives it",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderapi.Order"
                        }
                    },
                    "404": {
                        "description": "Order with ID 'X' not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/edit-order": {
            "patch": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Removes an order from the system",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Recipient",
                        "name": "recipient",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderapi.Order"
                        }
                    },
                    "404": {
                        "description": "Order with ID 'X' not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get-order": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Adds an order to the system",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderapi.Order"
                        }
                    },
                    "404": {
                        "description": "Order with ID 'X' not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/remove-order": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Removes an order from the system",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderapi.Order"
                        }
                    },
                    "404": {
                        "description": "Order with ID 'X' not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update-order-status": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Updates an order's status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "OrderRecieved",
                            "OrderProcessing",
                            "OrderOutForDelivery",
                            "OrderShipped"
                        ],
                        "type": "string",
                        "description": "Order Status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/orderapi.Order"
                        }
                    },
                    "404": {
                        "description": "Order with id 'X' not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "423": {
                        "description": "Order is no longer active",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "orderapi.IndexResponse": {
            "type": "object",
            "properties": {
                "documentationUrl": {
                    "type": "string"
                }
            }
        },
        "orderapi.Item": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "orderapi.Order": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/orderapi.Item"
                    }
                },
                "orderStatus": {
                    "$ref": "#/definitions/orderapi.Status"
                },
                "recipient": {
                    "type": "string"
                }
            }
        },
        "orderapi.Status": {
            "type": "string",
            "enum": [
                "OrderRecieved",
                "OrderProcessing",
                "OrderOutForDelivery",
                "OrderShipped"
            ],
            "x-enum-varnames": [
                "OrderRecieved",
                "OrderProcessing",
                "OrderOutForDelivery",
                "OrderShipped"
            ]
        }
    }
}